% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_column_from_dfs.R
\name{combine_column_from_dfs}
\alias{combine_column_from_dfs}
\title{combine_column_from_dfs}
\usage{
combine_column_from_dfs(dfs, key)
}
\arguments{
\item{dfs}{list of dataframes}

\item{key}{to extract column from df}
}
\value{
Vector of combined values of the column
}
\description{
combine_column_from_dfs
}
