% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tmp_line.R
\name{plot_tmp_line}
\alias{plot_tmp_line}
\title{plot_tmp_line}
\usage{
plot_tmp_line(df, x_value, y_value, color_value, end_value)
}
\arguments{
\item{df}{dataframe, source of plot}

\item{x_value}{column key used for x-axis}

\item{y_value}{column key used for y-axis}

\item{color_value}{column key used for group and color}
}
\value{
None
}
\description{
plot_tmp_line
}
