% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv.R
\name{read_csv}
\alias{read_csv}
\title{read_csv}
\usage{
read_csv(file_name, prefix = "")
}
\arguments{
\item{file_name}{file_name to load}

\item{prefix}{prefix for file path}
}
\value{
dataframe of loaded CSV
}
\description{
read_csv
}
