% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_over_dir.R
\name{read_csv_over_dir}
\alias{read_csv_over_dir}
\title{read_csv}
\usage{
read_csv_over_dir(dir_name)
}
\arguments{
\item{dir_name}{file_name to load}
}
\value{
list of dataframe of loaded CSV
}
\description{
read_csv
}
