% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_lm.R
\name{report_lm}
\alias{report_lm}
\title{report_lm}
\usage{
report_lm(data, x, y)
}
\arguments{
\item{x}{key of independent variable}

\item{y}{key of dependent variable}

\item{df}{dataframe, source of plot}
}
\value{
None
}
\description{
report_lm
}
