% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_norm.R
\name{report_norm}
\alias{report_norm}
\title{report_norm}
\usage{
report_norm(df, key)
}
\arguments{
\item{df}{dataframe, source of plot}

\item{key}{to get samples from dataframe and check normality}
}
\value{
None
}
\description{
report_norm
}
