% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_same_variance.R
\name{report_same_variance}
\alias{report_same_variance}
\title{report_same_variance}
\usage{
report_same_variance(data, x, y)
}
\arguments{
\item{x}{key to group samples and compare variance}

\item{y}{target variable when checking variance equality}

\item{df}{dataframe, source of plot}
}
\value{
None
}
\description{
report_same_variance
}
