% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_df_to_vec.R
\name{transform_df_to_vec}
\alias{transform_df_to_vec}
\title{transform_df_to_vec}
\usage{
transform_df_to_vec(df, key)
}
\arguments{
\item{df}{dataframe, source of plot}

\item{x}{key to get samples and build vector}
}
\value{
vector of target variable
}
\description{
transform_df_to_vec
}
